fix: remove systemd service only on package remove
On yum-based Linux systems, the package upgrade process installs the new
version first, then removes the old version. As a result, the update of HAS
on CentOS/RHEL results in a missing systemd service definition. This change
makes the removal of the systemd service dependent on the presence of the
"executable" that the service runs.