install/package: create perforce user and group
Similar to the p4d packages for Linux, the HAS Linux packages will create a
user and group named perforce as part of the installation process.

Since HAS includes an install shell script, an alternative to the packages,
that script will also create the perforce user and group. An option to not
create the user/group is available.