fix: SAML 1-step was missing a callback invocation
One of the possible code paths was missing an invocation of the callback
when preparing the response for the 1-step SAML authentication (as used by
Swarm). Unlikely to have ever surfaced as an issue, but nonetheless it
should be fixed.

Add more debug logging around the 1-step SAML request processing.

npm test passes