feat: validate JSON Web Tokens for authentication
Requires an issuer that can validate the JSON Web Tokens that it provides to
a trusted client. The /oauth/validate route introduced in this change is
used by the extension to verify the token provided by the users named in the
client-sso-groups and client-sso-users extension settings.

Introduces new settings (OAUTH_JWKS_URI, OAUTH_AUDIENCE, OAUTH_ISSUER,
OAUTH_ALGORITHM, and OAUTH_TENANT_ID) that are described in the example.env
file.

npm test passes