fix: expire cached data stored in Redis
The login requests and user profile data stored in Redis were not being
automatically purged after a set amount of time, unlike the in-memory data
cache when Redis is not in use. This change fixes that.

Additionally, the time-to-live for cached data can now be configured using
the CACHE_TTL setting, which is the expiration time of cached values in
seconds. Default is 300 seconds, or 5 minutes, which was the previous
default for the in-memory cache.

With this change, removed the 'transitory' dependency and switched to using
lokijs for in-memory cache for both the login and provisioning features.

npm test passes