feat: support Redis Sentinel with optional TLS
Support multiple Redis instances with automatic failover via Redis Sentinel.
See https://redis.io/topics/sentinel for instructions on setting up Redis
Sentinel. Support for TLS and non-TLS connections is provided, as well as an
optional password.

New configuration file, named by SENTINEL_CONFIG_FILE setting, which
configures the connection with the Redis Sentinel instances. An example is
given in sentinel.config.cjs with TLS for both the Redis instances as well
as the sentinels, and a password to connect to Sentinel. The minimum
configuation, with neither TLS nor passwords, is to give a `name` and the
list of `sentinels` (host name or IP address, port defaults to 26379).

npm test passes