feat: add status page indicating service status
Returned contents are JSON formatted for easier machine parsing. Checks if
the certificates are valid and not expiring soon, if the OIDC issuer
responds with a valid configuration, if the SAML metadata URL returns valid
metadata, if the Redis server responds appropriately to 'ping', and if the
parameters to connect to Helix Core Server indicate a valid login ticket. If
everything is either 'ok' or 'not configured', then the overall reported
status is 'ok'.

Does not attempt to validate SAML when configured from a metadata file or an
IdP SSO URL. These are not useful for knowing if the SAML IdP is responsive
or not, which is the primary concern with the server status. Whether the
SAML configuration is valid or not can only be tested manually during
installation.

npm test passes