build: produce docker image for builds on CentOS 8
Because the official CentOS 8 package repository has gone offline, and the
replacement repository is extremely unreliable, the builds and tests now use
a docker image that has been pre-built and copied to the build hosts. By
building and testing using this base image we should avoid much of the issue
caused by CentOS 8 going into retirement.