feat: allow acsUrlRe in SAML IdP configuration
If the acsUrl is not found in the configuration file named by the
IDP_CONFIG_FILE setting, and a property named acsUrlRe _does_ exist, then
use that and treat it as a regular expression. This is to support Swarm when
configured with more than one Helix Core Server instance, in which the ACS
URL that Swarm constructs contains an additional path entry.

For instance, if Swarm is configured with only a single instance, then the
acsUrl might look like this:

https://swarm.example.com/api/v10/session

However, when Swarm is configured to connect with multiple servers, then the
ACS URL will have a server identifier, like so:

https://swarm.example.com/chicago/api/v10/session

Such a URL would not match the static acsUrl defined in the IDP_CONFIG_FILE.
To allow for multiple different ACS URL values, the service now looks for a
property named acsUrlRe, if acsUrl is not present. This value is then
treated as regular expression to match against the ACS URL provided by
Swarm. Such a regular expression might look like this:

https://swarm.example.com/[^/]+/api/v10/session

undoc: for now this feature is not documented.

npm test passes