test: verify OAuth token validation is secure
Technically there was a coding mistake in the arguments passed to the JWT
verification function (should have been `algorithms` but was `algorithm`).
While the verifier was unaffected by this mistake, in that the verification
was not fooled by token forgery, it was still a mistake. The tests now
demonstrate that the verification is working as expected, despite attempts
to forge the token.

npm test passes