feat: allow tighter JSON web token verification
Allow the administrator to explicitly set which JWKS key identifier is
allowed, rather than trusting the `kid` value in the JWT header. This is
done with the new setting OAUTH_JWKS_KEYID which defaults to
undefined (trusting the JWT kid value).

Further, the administrator can set the public key rather than using the both
`kid` value and the JWKS URI by setting the OAUTH_SIGNING_KEY_FILE value.
Setting this makes both the OAUTH_JWKS_URI and OAUTH_JWKS_KEYID settings
irrelevant, as the service will verify the token from the signing key alone.

Both of these new settings are optional; the JWT verification is already
difficult to circumvent, and assuming the JWKS URI is using HTTPS, the
chances of a MITM exploit are slim.

npm test passes