build: use rockylinux:8 vs unreliable CentOS 8
The CentOS 8 repositories are extremely unreliable to the point that the
builds are failing nearly all of the time. The only solution is to switch to
something else as it has proven to be very difficult to prevent yum from
contacting the remote servers. The freely available RockyLinux seems to be
the practical and spiritual successor of CentOS, and as such is now the
basis of our builds for CentOS 8.

Note that the CentOS 7 builds are unchanged, using the CentOS 7 image.