feat: allow setting OIDC token signing algorithm
In very rare circumstances it is helpful to be able to configure HAS to
expect the OpenID Connect ID token to be signed using an algorithm other
than the default RS256. While that value works with the vast majority of
identity providers, it is possible for an admin to configure (PingFederate)
to use a different algorithm, which causes HAS to reject the response.

With this change a new setting named OIDC_TOKEN_SIGNING_ALGO is introduced,
which if defined, will be passed to the OpenID client library. The default
value is RS256, which has been the default since the beginning.

npm test passes