build: try building CentOS package on centos:8
Using rockylinux:8 maybe contributed to the bug in which yum fails to
install the package because it attempts to connect to "http://media:/" which
is horribly wrong in so many ways. No idea where this value is coming from,
it is not in the package itself, and installing using `rpm -i` works fine.

The drawback of using the centos:8 docker image is that the builds will most
likely fail most of the time, causing a lot of grief. May have to insert a
loop in the build procedure to keep trying until the build passes.