fix: prohibit root user running configure script
Allowing the root user to run the configure script creates issues like the
one where client-secret.txt is not readable by the process that starts the
service. Typically the service starts and runs as the "perforce" user, but
if the client-secret.txt is owned and readable only by root, then the
service will fail to process OIDC requests.

This change now blocks the root user from running the configure script, in
the same manner as the install script. Both scripts now have an --allow-root
option for those users who can work-around any file permissions issues that
result from running the scripts as the root user.