fix: read configuration files with old extension
If the service is configured with LOGGING or IDP_CONFIG_FILE settings that
point to configuration files with the CommonJS format and yet have the .js
extension, then automatically convert the contents in-memory to allow
Node.js to read those configuration files correctly. Normally with an ES6
module, Node.js treats any .js file as having ES6 format, but these
configuration files still use the CommonJS format, and hence are expected to
have the .cjs extension. However, in an upgrade scenario, this will not be
the case. As such, HAS now detects this at runtime and corrects the format
automatically, without modifying the files.

npm test passes