fix: start service as root then switch user/group
The service cannot bind to a low-numbered port (<1024) when it is running as
an unprivileged user. Thus, the service will now start as the root user,
like before, and then switch the user and/or group of the process after
binding to the desired port. The user and group are defined by the new
settings SVC_USER and SVC_GROUP, if provided. They can be either the name of
a user and group, or the numeric identifier.

npm test passes