fix: config script to set file ownership correctly
The configure script was creating and/or changing the ownership of the files
it touched to whatever ${USER} happened to be. If the script is invoked
using `sudo` then the USER is root so the files end up owned by root. A
better approach is to use `chown --reference` to copy the ownership from
another file, in this case example.env which we expect to exist.