fix: prevent hanging login with concurrent logins
If the user were to start a login with p4 and then cancel it with Ctrl-c,
then start another login attempt immediately after, the second attempt would
hang until the login timeout period ended. Because the first login left the
extension running, that first invocation retrieved the user profile data
from the service, while the second invocation was left waiting for the data
that would never come again. This is due to the service clearing the user
profile data when the first request retrieves it (to prevent replay attacks).

To correct this behavior, the user profile data is now keyed by the unique
request identifier, rather than the user identifier provided by the client
application. There was never a good reason for this in the first place, it
was merely an oversight. Most of this change is fairly trivial, but the one
significant change is the addition of two new use cases, one for generating
the SAML response to Swarm and another for validating that response when it
is sent from the extension. This is necessary in order for the request
identifier to be carried from the user login session to the request made by
the extension, which is not associated with the user's browser session.

npm test passes