feat: logout and reset password for inactive users
When configured for user provisioning support, and a PATCH or PUT on a user
entity is received in which the 'active' property is false, then that user
will be logged out of Perforce and their password set to a long random
string. If either the logout or password change fails for any reason (there
are at least a few reasons why either might fail), the errors are ignored.

docs: after a user is made inactive, they will not be able to authenticate
with Perforce again until an admin changes their password. Ideally the cloud
service provider would push the user update with a new password. In the SSO
scenario, the use would simply be re-enabled in the identity provider.

npm test passes