feat: allow key pinning of client certificate
The client certificate used to retrieve the user profile data can be
restricted by the CLIENT_CERT_CN setting in HAS, which compares the common
name of the certificate to a given pattern. This works, but an even better
approach would be to compare the digest, or fingerprint, of the certificate
to an expected value. This change introduces the CLIENT_CERT_FP setting
which takes the SHA256 digest of the client certificate (with colons
separating each byte, as in AA:BB:CC:DD:...). This is more secure and in
line with common practice in terms of key pinning.

npm test passes