feat: option to allow OIDC user to select account
New setting OIDC_SELECT_ACCOUNT can be set to true to prompt the user to
select an account when authenticating with an OIDC identity provider.
Without this setting, the default behavior is as before, no prompt is
requested by HAS during the login. The IdP is always free to do whatever it
deems appropriate, this setting is only a request by the auth service.

npm test passes