feat: login test URL shows user profile details
A new `loginTestUrl` provides a means for administrators to not only test
the authentication process but also to see the full user profile data
returned by the identity provider in the browser. This is more convenient
than finding the log file and scanning through its contents, and also helps
when configuring the p4d extension.

npm test passes