fix: override node-forge dependency to fix vuln
The node-forge package is indirectly required via another dependency, and
that version has a vulnerability (CVE-2022-0122). To temporarily fix this,
override the node-forge version with a fixed version.

Note that this does not address development related vulnerabilities as those
cannot be overridden, and only affect the unit tests, not the production
code.

npm test passes