chore: switch to scoped node-saml, passport-saml
The node-saml and passport-saml modules have been (temporarily) scoped under
the @node-saml namespace, and as such the references to those dependencies
has been updated. At some point both packages will be released properly and
the scoping will no longer be necessary, probably.

As a result of this change, the versions have been updated and some of the
checks are more stringent. In particular, the audience in the SAML response
is now validated against expectations. Since we can have multiple audiences
thanks to the flexibility of the contents of the IDP_CONFIG_FILE, this is
not an easy fix. For now, the audience is set to `false` to disable that
particular check, which continues its previous behavior.

npm test passes