admin: basic support for service address changes
The REST API and the web interface now provide basic support for the address
of the service changing during a settings update. That is, the REST API will
respond to /settings/apply with the Location of the service, in case it was
changed (via SVC_BASE_URI, for instance). The web interface will similarly
show the new location when settings are applied, allowing the admin to click
a link to go to the new address.

The web interface cannot reliably send the browser to the new address as it
will fail completely when running in development mode, and there seems to be
no means for knowing whether the web client is running in a development or
production environment.

npm test passes