config: prompt to configure admin web interface
The configure script now prompts the user with the choice of setting up the
administrative web interface. Also fixed a bug in which the configure script
would read the existing .env file and set values as SETTING=SETTING in some
cases. Added a docker setup for manual testing of the configure script.

npm test passes