fix: assert the idp cert is provided for SAML
When preparing the SAML strategy for passport, if the `cert` property is
missing, raise a clear and obvious error to the user (and in the logs). It
became required by the node-saml library some time ago that the IdP have a
public certificate in the SAML configuration. By the time the service has
processed its configuration, along with the optional IdP metadata, there
should be a cert property. If not, it is treated as an error.

npm test passes