fix: prompt for SAML SP entityId unconditionally
The configure script was skipping the prompt for the SAML SP entityId if the
metadata URL or SSO URL were not provided. However, if the admin was
planning to use the metata FILE property instead of the URL, then the script
would skip the prompt for a necessary piece of information.