fix: report missing P4PASSWD as an error
The SCIM feature cannot function without P4PORT, P4USER, and P4PASSWD. As
such, an error is now raised when P4PASSWD is not set when attempting to use
the user provisioningg feature. If using the configure script, this should
not be the case, but if the .env file is edited manually, then it is
possible that only some of the settings are specified.

npm test passes