test: add tests for /multi/login route
Minor fix to the error reporting that was rarely seen or tested, and as such
there was some mishandling of view "local" variables.

Also, the read/write config usecases now handle AUTH_PROVIDERS(_FILE) in a
manner that is consistent with other file-based settings, in preparation for
supporting the administrative interface supporting multiple providers.

npm test passes