fix: without request id, show special login page
Rather than redirecting to the landing page when the browser session is
missing the login request identifier, instead show a page to indicate that
the login was only partially successful. This is also useful when testing
the IdP-initiated login scenario, as that sends a SAML authentication
response to the service when it was not expecting one.

npm test passes