fix: perform user logout when being made inactive
When a user is made inactive, they are logged out of Helix Core Server and
their password is changed to a random string. However, this was not working
when the cloud service provider used the "external" identifier for the
user, which includes a "user-" prefix to distinquish from groups.

npm test passes