fix: address CVE-2022-39299 with updated libraries
Update to the "beta" tagged beta versions of the node-saml and passport-saml
libraries to get the fix for CVE-2022-39299. This introduces a breaking
change in that the SAML response and assertion are expected to both be
signed, where previously either one was sufficient.

Two new service settings have been introduced:

SAML_WANT_ASSERTION_SIGNED, defaults to true, but can be set to 'false' to
disable the check for a signed assertion.

SAML_WANT_RESPONSE_SIGNED, defaults to true, but can be set to 'false' to
disable the check for a signed response.

Note that the node-saml library will refuse any response for which neither
the response nor the assertion have been signed, regardless of settings.

Auth0 will either sign the assertion (its default) or sign the response, but
not both. As such, setting SAML_WANT_RESPONSE_SIGNED=false in the service
configuration is required. Or, change Auth0 to sign the entire response and
then set SAML_WANT_ASSERTION_SIGNED=false in the service configuration.

AzureAD defaults to signing the assertion, so either change its
configuration to sign both, or add SAML_WANT_RESPONSE_SIGNED=false to the
service configuration.

Google Workspace defaults to signing the assertion, and provides an option
to sign the response, so either enable that option or add
SAML_WANT_RESPONSE_SIGNED=false to the service configuration.

Okta requires no change as it defaults to signing both the assertion and the
response.

OneLogin will default to signing the response, so either change its
configuration to sign both, or add SAML_WANT_ASSERTION_SIGNED=false to the
service configruation.

npm test passes