fix: fix group provisioning for empty groups
Some SCIM clients seem to perform multiple updates with seemingly no change
at all. In particular, the same empty group will be PUT twice, which
resulted in an error in our SCIM endpoint. This has been fixed. Also, the
adding of a user has been made consistent with updating, to eliminate the
chance of their being any errors from strange requests (like sending a new
user with our user ID prefix already prepended). And yet another fix is to
handle the user `id` value consistently in the HTTP response, where
previously it would be the display name, which is not a good identifier.

npm test passes