fix: return consistent userName for user queries
The result when retrieving a single user versus retrieving mulitple users
was slightly different, in that the userName returned might not match
expectations. In particular, if the userName was originally not a name that
conformed to Perforce username standards (i.e. it was an email address),
then the original userName would be saved in the Perforce key/value table.
When retrieving a single user record, however, that value was ignored and
the value used to retrieve the record was returned, while the multi-user
fetch would use the key values.

npm test passes