fix: return externalId and active user properties
Previously certain SCIM clients would provision and deprovision users as
expected, but recently they appear to expect different behavior than what
was previously released. In particular, the externalId is not optional in
their eyes, and therefore must be returned. Similarly, not including the
active field was interpreted as being inactive.

npm test passes