fix: retain existing SAML validation behavior
After updating the node-saml library to the latest (beta of beta) version,
both the SAML response and assertion were required to be signed by the
identity provider. Most providers do not sign both by default, and at least
one of them cannot sign both. With this change, either the response or the
assertion must be signed, but having both signed is not a requirement. This
will avoid breaking existing installations on upgrade, causing much grief
for users when attempting to authenticate via the service. In detail, the
two settings SAML_WANT_RESPONSE_SIGNED and SAML_WANT_ASSERTION_SIGNED will
both default to 'false' instead of 'true'; note that regardless of these
settings, either the response or the assertion _must_ be signed as the
node-saml library will not accept anything less.

Thomas assisted with the wording in the updated RELNOTES which addresses
this change, and how the next release will see the defaults for the two
aforementioned settings reverting back to 'true' which will fully address
the vulnerability in question (CVE-2022-39299).

npm test passes