Default config values defined in a single file
All of the configuration settings that have defaults values that are not
computed based on other settings (such as DEFAULT_PROTOCOL) are now defined
in a file called defaults.env located at the base of the installation path.
The administrative interface now returns all of these settings as if they
were defined in the environment, yielding more consistent results between
the backend environment and frontend form values.

npm test passes