test: test for Node.js v14 and 16 compatibility
Using the magic of docker, run the true unit tests (those that do not rely
on external services and binaries) in an environment with an older release
of Node.js installed. This set of environments is to include only those
versions of Node.js that are in "maintenance", which presently includes v14
and v16.

npm test passes