Remove install/config support for pm2 process manager
Questions have been raised externally about the legal implications of using
the pm2 process manager, and since moving to systemd has proven to be
overall a better option, pm2 is going away. The install and configure
scripts no longer have any knowledge or support for pm2, and the example
configuration file has been removed.

This change also fixes the tests that were broken as a result of the change
from using Node 16 to 18. Not all Linux releases have readily available
support for Node 18, so the install script and packages will now install the
appropriate version for each specific OS release.