fix: API response to "not found" should be terse
The response to /requests/status for a request identifier that did not exist
would return a large blob of HTML. This often ends up in the log files of
the client application, such as the Helix Core extension. Instead, the
response should be as brief as possible, and is now "no such request".

npm test passes