fix: support separate cert/key pair for SAML/Swarm
In cases where a customer employs a service like Let's Encrypt, the service
key and cert pair will change on a regular basis (every 2-3 months). This
will cause problems for integration with a SAML IdP that expects the service
to sign authentication requests, or with the Swarm integration that
absolutely must have the correct service certificate.

To prevent this problem for customers that frequently rotate the service
certificates, a separate cert/key pair can be employed exclusively for the
SAML IdP and Swarm integration. The settings for this are named
SAML_CERT_FILE and SAML_KEY_FILE, and if present, take precedence over the
usual CERT_FILE and KEY_FILE settings. If the new settings are not defined,
then the CERT_FILE and KEY_FILE will be used, as before.

npm test passes