feat: prevent phishing by seeking user approval
In order to prevent phishing attacks in which an unsuspecting user is
tricked into clicking a login URL that authenticates a login request, an
option is now available that forces all users to authorize the login request
every single time. Simply set PROMPT_FOR_AUTHORIZATION to any value in the
.env file and HAS will display a web page informing the user that a login
request is being made, presenting them with a form/button to click in order
to approve the login and proceed to the configured identity provider.

npm test passes