fix: show dotenv results during startup and reload
After the recent changes related to the refreshing of the environment
configuration, the "dotenv results" log entry would only appear when the
configuration was refreshed, but not during startup. Previously the output
would be shown during startup. Now it will appear both during startup and
when the configuration is refreshed.

npm test passes