fix: minimize error for missing .env file
It is not an error if the .env file is missing, it simply means the
administrator configured the service using environment variables using other
means. This is especially true with the Docker Hub image, in which we
suggest starting the container using Docker Compose and its 'env' section of
the compose file.

The error message (not stack track) is still logged at debug level, in case
a missing .env file is actually not expected.

npm test passes