fix: allow P4PASSWD to be a ticket value
The documentation says that for the P4PASSWD setting in the service, a
ticket value is permitted. However, this was not working until now. If the
password value appears to be a ticket, and the connection is already
authenticated, then do not perform a login as it would fail anyway.

npm test passes