doc: fix examples for default protocol in docker
Recently changed the docker build such that the SVC_BASE_URI was removed and
PROTOCOL was set to 'http'. However, the README for the image on Docker Hub
provided examples that did not work out of the box since the SVC_BASE_URI
started with https: but the protocol was http. Fixed the examples and added
a note about setting the PROTOCOL apporpriately. The basic intent is that
the service will be run from behind a reverse proxy that performs the TLS
termination, as that seems more appropriate when using a docker host.