fix: include optional module to fix syslog logging
Having the service log using the syslog facility was failing (when
installing the service using the prebuilt packages) because the optional
dependency unix-dgram was missing. This module is used by winston-syslog
in order to log to syslog.

Also fixed the package build steps to avoid copying the very large client
node_modules directory.