feat: IdP metadata text area added to admin ui
Some identity providers, such as Google Workspace, do not offer a URL for
retrieving the SAML metadata. Instead, the XML must be copied and pasted to
a file for the authentication service to read. In this change, a text area
is added to the admin interface that will accept the metadata text, saving
it to a file on the server automatically (much like how the certificates are
handled).