feat: ability to change configuration temporarily
Add a new REST API endpoint for defining configuration changes in a
temporary fashion, with a second new endpoint for removing the temporary
changes. This can be used to try changes to the configuration without
committing them to the configuration file. Reverting the changes is easier
this way, as the changes are stored in memory and not written to the
configuration file. If the service is restarted, the temporary changes are
lost.

The changes to the admin ui are temporary as well, since the UX team is
working on a major redesign.

Many of the changes are related to changing the constructor of
MapSettingsRepository, which affected many of the unit tests. The actual
changes to the service itself were minor.

npm test passes