test: make shibboleth dependent on healthy ldap
Using `depends_on` alone is not sufficient for the shibboleth container to
start reliably. Its startup, along with the ldap container, is timing
dependent and thus can fail if LDAP is not immediately responsive. This
change makes the depends_on dependent on the health of the ldap container,
not just that it was instructed to start.