fix: use DEL instead of UNLINK Redis command
If the Redis server is a version prior to 4.0 then the service will silently
fail to remove user objects from the cache (note that they will eventually
expire on their own). This presents a window of opportunity for an attacker
to get the user profile data if they have access to HAS or Redis and know
the user identifier, and the customer is using an older version of Redis.

This change replaces `unlink` with `del` which has the same basic effect and
is compatible with all versions of Redis since 1.0.

npm test passes