fix: log error and hang around indefinitely
It is not currently possible to write to a file-based winston log and then
immediately exit. Nor do any of the "graceful" methods for exiting work as
the flush of the logs is broken in winston. As a result, the best that can
be done for now is to log the error and stay running until the admin
realizes the problem and fixes the configuration.

npm test passes