test: configure containers to work with cgroup v2
The Linux kernel has had support for cgroup v2 since 2016 and that has now
made its way into both systemd and docker. With docker 20.10 (current
version is 23.0.1) the use of cgroup v2 has been made the default. This
affects how containers running systemd should be started, namely the need to
map in the cgroup fs is no longer necessary (and in fact will cause the
container to fail to start properly). The containers used for testing the
install script and package installations now employ the updated usage.