fix: treat FORCE_AUTHN as with previous releases
In releases prior to 2023.1, the FORCE_AUTHN setting merely had to be set to
some value and then user authentication would be required on every request.
However, during development for 2023.1, the behavior was mistakenly changed
such that FORCE_AUTHN was treated as a strict true/false boolean value.

In addition to restoring the previous behavior, found and fixed a similar
issue with boolean-like SAML settings (SAML_DISABLE_CONTEXT, SAML_WANT_*).

npm test passes