chore: switch from create-react-app to Vite
The old create-react-app (CRA) is effectively retired (it has 400+ open pull
requests and few recent commits). The current advice is to use Vite ("veet")
which is similar to CRA but much faster. As a result the client application
is now built using Vite instead of CRA. This has no impact on users but does
change the commands used to start and build the frontend. See the README.md
in the client directory for details.