chore: drop support for Node.js v14, allow v20
While v20 is not yet available, it should work in theory, and as such the
install and configure scripts will allow for it. Meanwhile, v14 is EOL by
the end of the month, so removing support for that. As a result, several
dependencies have been updated as they did not support v14.

npm test passes