feat: REST API for validating Swarm configuration
If the VALIDATE_ENABLED environment variable is set to 'true' when the
service is started, the /validate/swarm route will be enabled. A Swarm
config.php file can be uploaded via POST using multipart/form-data and a
JSON response will indicate if there are any issues.

The response will contain a property named 'status' that will have one of
several values. If 'ok', then there are no problems. If 'errors' then the
property named 'errors' will be a list of lexer/parser errors (meaning that
the configuration file was malformed). Lastly, status could be 'results'
which means the 'results' property will have a list of configuration issues.

npm test passes