Unfinished rewrite of the admin interface
In an effort to support the new administrative interface, which is geared
around the adding and editing of multiple identity providers, some nearly
imperceptible changes were made to the backend. Internally the classic OIDC
and SAML settings are converted into the same set of "auth providers" that
is exposed via the REST API. That is, the application code now defers to the
GetAuthProviders usecase to process the settings and return a set of
identity providers, rather than reading individual OIDC or SAML settings.
When settings are changed via the REST API and written back to the .env
file, they will be flattened back into the OIDC/SAML settings, if and only
if, there is exactly one OIDC and/or one SAML provider configured.

Not everything is working as expected yet, and the admin interface is still
incomplete, but the basic add and edit functions are working. For instance,
adding another OIDC or SAML provider will result in the addition of the
AUTH_PROVIDERS setting, but the old OIDC and SAML settings are still
present. This won't matter to the application, but it leaves clutter in the
configuration file and leads to confusion.

npm test passes