fix: fix multiple providers support
Firstly, when writing the providers to AUTH_PROVIDERS, remove the id
property as that field is meant to be assigned at runtime. It is vital that
the id be assigned in a predictably consistent manner that will result in
the same values across multiple instances of the service.

Second, when producing the list of auth providers, always tidy the list so
that the incomplete providers resulting from the default settings do not
appear in the list.

npm test passes