fix: manage default settings within providers
Previously, when hydrating providers from classic settings, the default
settings would be incorporated, but when processing the providers defined in
the AUTH_PROVIDERS list, the defaults were not injected. This change fixes
that. Additionally, when writing the AUTH_PROVIDERS value to the .env file,
the defaults are stripped, just as with the classic settings. This should
reduce the clutter and ensure consistent behavior regardless of how the
providers are defined in the configuration.

npm test passes