bug: forceAuthn for a provider now works properly
While multiple providers was available earlier, it was not possible for a
client application (e.g. p4d extension) get get a login URL for a specific
provider (it was always up to the user to choose the provider during the
login process). With the upcoming release, that is now possible by adding
the providerId query parameter (this is used by the admin ui to test login
for a specific provider). The docs state that forceAuthn can be set on a
per-provider basis, but this was not really effective until now. As such,
this issue was not really exposed previously, but nonetheless it was not
really working either.

npm test passes