admin: add text area for IdP metadata and cert
For Google Workspace (SAML), need a field for putting in the IdP certificate
as they do not provide a metadata URL (or metadata in any form). May as well
also add a text area along side the certificate for entering the raw IdP
metadata, if the customer chooses that over using the URL.

At this point, all providers should be supported and working.