fix: add more error reporting for p4 entity repo
The Helix entity repository implementation was not checking for errors in
some cases, such as invoking `p4 key`, and that would result in an error if
the p4d response was not as expected. Now, more error reporting has been
added so that whatever error p4d is returning will get reported to the user
via the log file.

npm test passes