admin: error page "go back" will refresh properly
When an error occurs on the home screen of the admin interface, the "go
back" button would do nothing. This was because nagivate("/") had no effect
when the route was already "/". As such, if the location is already "/" then
the "go back" button will cause the page to reload instead.