doc: example of NGINX and HAS with provisioning
When HAS is configured for user provisioning, it must be accessible to the
public internet, as well as have a working connection to Helix Core Server.
This example shows how to use NGINX as a reverse proxy, forwarding the
Authorization header to HAS for the /scim/v2 path. This allows for HAS to be
hidden behind a firewall and reduce the inherent security risk.