fix: prefer BEARER_TOKEN_FILE over BEARER_TOKEN
Exposing the bearer token in the .env configuration file is a much worse
scenario than just about anything else as it is the only "secret" that
controls access to the Helix Core Server users and groups. As such, the
configure script will now write the provided bearer token to a new file and
use the new BEARER_TOKEN_FILE setting.

npm test passes