fix: write p4 ticket as P4PASSWD in .env file
Rather than writing the super user password directly to the .env file, the
configure script will now write the p4 ticket value instead. This acts as a
marker that the service will detect and instead of invoking `p4 login` with
the provided password, it will invoke `p4 login -s` to check the login
status. This prevents exposing the super user password when the customer
sends their .env file to support.

Writing the p4 credentials to the .env file is engaged when the user chooses
to configure the service for user provisioning. Otherwise these values are
not collected nor written to the .env file.