fix: treat authnContext as a list in REST API
Previously the REST API would return the SAML provider's authnContxet
property as string, even if it contained multiple values separated by
commas. Worse, if the values were wrapped in quotes and the entire string
contained square brackets ([]) then the administrative interface would show
this value as-is in the "Authentication Context" text field when editing the
SAML auth provider.

Now the REST API will convert the string to an array and the client can then
work with the individual values in whatever manner is appropriate. In the
case of the web interface it simply joins them with commas, minus the quotes
and square brackets.

npm test passes